+++ packages/data-mate/src/index.ts
@@ -88,5 +88,5 @@ export {
 export * from './adapters/index.js';
 export * from './function-configs/index.js';
 
-export { makeRandomDataSet } from './radom-DT-data-generator/make-dt-data-set.js';
-export { makeRandomDataFunctionForField } from './radom-DT-data-generator/make-dt-field-fn.js';
+export { makeRandomDataSet } from './random-DT-data-generator/make-dt-data-set.js';
+export { makeRandomDataFunctionForField } from './random-DT-data-generator/make-dt-field-fn.js';
